refactor: split percentile code in separate file
